initiating awyis

+++ server/server.js
@@ -16,23 +16,28 @@ app.listen(3010, ()=> {
 //                    AXIOS                      //
 // ============================================= //
 
-axios.get('https://mhw-db.com/skills').then((Awyis) => {        
-    fs.writeFile('./server/JSON/MHW-skills.txt', JSON.stringify(Awyis.data, null, 2), (err) => {
-        if (err) return res.json({error: 'cannot write MHW-skills file'})
-        console.log("MHW-skills file has been created")
-    })
-}).catch((err)=> {
-    return res.json({error: err})
-})
 
-axios.get('https://mhw-db.com/armor').then((Awyis) => {
-    fs.writeFile('./server/JSON/MHW-armor.txt', JSON.stringify(Awyis.data, null, 2), (err) => {
-        if (err) return res.json({error: 'cannot write MHW-armor file'})
-        console.log("MHW-armor file has been created")
-    })
-}).catch((err)=> {
+if (fs.existsSync('./server/JSON/MHW-skills.txt') === false) {
+    axios.get('https://mhw-db.com/skills').then((Awyis) => {        
+        fs.writeFile('./server/JSON/MHW-skills.txt', JSON.stringify(Awyis.data, null, 2), (err) => {
+            if (err) return res.json({error: 'cannot write MHW-skills file'})
+            console.log("MHW-skills file has been created")
+        })
+    }).catch((err)=> {
     return res.json({error: err})
-})
+    })  
+}
+
+if (fs.existsSync('./server/JSON/MHW-armor.txt') === false) {
+    axios.get('https://mhw-db.com/armor').then((Awyis) => {
+        fs.writeFile('./server/JSON/MHW-armor.txt', JSON.stringify(Awyis.data, null, 2), (err) => {
+            if (err) return res.json({error: 'cannot write MHW-armor file'})
+            console.log("MHW-armor file has been created")
+        })
+    }).catch((err)=> {
+        return res.json({error: err})
+    })
+}
 
 // ============================================= //
 //                    ROUTER                     //
@@ -44,5 +49,14 @@ app.get('/api/skills', (req, res)=> {
         
         let skillsData = JSON.parse(data)
         return res.status(200).json(skillsData)
+    });
+})
+
+app.get('/api/armor', (req, res)=> {
+    fs.readFile('./server/JSON/MHW-armor.txt', (err, data) => {
+        if (err) throw res.json({error: 'cannot read armor file'});
+        
+        let armorData = JSON.parse(data)
+        return res.status(200).json(armorData)
     });
 })
